:recycle: clean up global search views
