+++ aksel.nav.no/website/app/_ui/global-search/GlobalSearch.provider.tsx
@@ -1,74 +1,107 @@
 "use client";
 
 import {
-  createContext,
-  useContext,
+  useCallback,
   useEffect,
   useMemo,
   useRef,
   useState,
+  useTransition,
 } from "react";
+import { debounce } from "@navikt/ds-react";
+import {
+  type GlobalSearchActionReturnT,
+  GlobalSearchResultContext,
+  useGlobalSearch,
+} from "@/app/_ui/global-search/GlobalSearch.context";
+import {
+  fuseGlobalSearch,
+  preloadSearchIndex,
+} from "@/app/_ui/global-search/server/GlobalSearch.actions";
+import { useParamState } from "@/app/_ui/global-search/useParamState";
+import { umamiTrack } from "@/app/_ui/umami/Umami.track";
 
-type SearchContextType = {
-  open: boolean;
-  openSearch: () => void;
-  closeSearch: () => void;
-  inputRef: React.MutableRefObject<HTMLInputElement | null>;
-};
+function GlobalSearchResultProvider({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  const { openSearch } = useGlobalSearch();
+  const shouldInitialOpenRef = useRef<boolean>(true);
 
-const SearchContext = createContext<SearchContextType | null>(null);
+  const [searchResult, setSearchResults] =
+    useState<GlobalSearchActionReturnT | null>(null);
 
-function GlobalSearchProvider({ children }: { children: React.ReactNode }) {
-  const inputRef = useRef<HTMLInputElement | null>(null);
-  const [open, setOpen] = useState<boolean>(false);
+  const [, startTransition] = useTransition();
+  const { setParam, clearParam, paramValue } = useParamState("query");
+
+  const debouncedSearch = useMemo(
+    () =>
+      debounce((query: string) => {
+        maybeEnableComicSans(query);
+
+        umamiTrack("sok", {});
+        setParam(query);
+      }, 200),
+    [setParam],
+  );
+
+  const resetSearch = useCallback(() => {
+    debouncedSearch.clear();
+    clearParam();
+  }, [clearParam, debouncedSearch]);
 
   useEffect(() => {
-    const listener = (event: KeyboardEvent) => {
-      if (
-        (event.key === "k" || event.key === "b") &&
-        (event.metaKey || event.ctrlKey) &&
-        !event.shiftKey
-      ) {
-        event.preventDefault();
-        if (open) {
-          inputRef.current?.select();
-        } else {
-          setOpen(true);
-        }
+    if (!paramValue) {
+      setSearchResults(null);
+      shouldInitialOpenRef.current = false;
+
+      return;
+    }
+
+    startTransition(async () => {
+      const newResults = await fuseGlobalSearch(paramValue);
+
+      setSearchResults(newResults);
+
+      /*
+       * We use a ref to ensure the dialog only auto-opens after results are in to avoid CLS.
+       */
+      if (shouldInitialOpenRef.current) {
+        openSearch();
+        shouldInitialOpenRef.current = false;
       }
-    };
+    });
+  }, [paramValue, openSearch]);
 
-    document.addEventListener("keydown", listener);
-
-    return () => document.removeEventListener("keydown", listener);
-  }, [open]);
+  /**
+   * Preload the searchindex cache, so that the first search is faster.
+   *
+   */
+  useEffect(() => {
+    void preloadSearchIndex();
+  }, []);
 
   const contextValue = useMemo(
     () => ({
-      open,
-      closeSearch: () => setOpen(false),
-      openSearch: () => setOpen(true),
-      inputRef,
+      queryResults: searchResult,
+      updateQuery: debouncedSearch,
+      resetSearch,
     }),
-    [open],
+    [searchResult, debouncedSearch, resetSearch],
   );
 
   return (
-    <SearchContext.Provider value={contextValue}>
+    <GlobalSearchResultContext.Provider value={contextValue}>
       {children}
-    </SearchContext.Provider>
+    </GlobalSearchResultContext.Provider>
   );
 }
 
-function useGlobalSearch() {
-  const context = useContext(SearchContext);
-  if (!context) {
-    throw new Error(
-      "useGlobalSearch must be used within a GlobalSearchProvider",
-    );
+function maybeEnableComicSans(query: string) {
+  if (query.includes("comic")) {
+    document.body.style.fontFamily = "Comic Sans MS, cursive, sans-serif";
   }
-
-  return context;
 }
 
-export { GlobalSearchProvider, useGlobalSearch };
+export { GlobalSearchResultProvider };
