+++ aksel.nav.no/website/app/_ui/global-search/server/GlobalSearch.config.ts
